requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
fastapi==0.135.3
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.43.0
watchfiles==1.1.1
websockets==16.0
